+++ plone/formwidget/masterselect/interfaces.py
@@ -1,8 +1,9 @@
+from z3c.form.i18n import MessageFactory as _
 from zope.interface import Interface
 from zope.schema import Tuple
-from zope.schema.interfaces import IObject
-
-from z3c.form.i18n import MessageFactory as _
+from zope.schema.interfaces import IBool
+from zope.schema.interfaces import IChoice
+from zope.schema.interfaces import IField
 
 
 class IMasterSelectWidget(Interface):
@@ -20,24 +21,32 @@ class IMasterSelectRadioWidget(Interface):
     """
 
 
-class IMasterSelectField(IObject):
+class IMasterSelectField(IField):
     """
     Additional Fields for MasterSelect
     """
     slave_fields = Tuple(
-        title=_(u"title_slave_fields", default=u"Fields controlled by this field,"),
-        description=_(u"description_slave_fields", default=u"Fields controlled by this field, if control_type "
-                      "# is vocabulary only the first entry is used"),
+        title=_(
+            "title_slave_fields",
+            default="Fields controlled by this field,",
+        ),
+        description=_(
+            "description_slave_fields",
+            default="Fields controlled by this field, if control_type "
+            "# is vocabulary only the first entry is used",
+        ),
         default=(),
-        required=False
+        required=False,
     )
 
-class IMasterSelectBoolField(IMasterSelectField):
+
+class IMasterSelectBoolField(IMasterSelectField, IBool):
     """
     Additional Fields for MasterSelect
     """
 
-class IMasterSelectRadioField(IMasterSelectField):
+
+class IMasterSelectRadioField(IMasterSelectField, IChoice):
     """
     MasterSelect radio button widget
     """